pyproject.toml
[tool.poetry]
name = "issarvideos"
version = "0.1.0"
description = ""
authors = ["nirkon <nirk@originai.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
requests = "^2.32.3"
pytest = "^8.3.5"
openai = "^1.84.0"
python-dotenv = "^1.1.0"
selenium = "^4.33.0"
google-auth = "^2.40.3"
google-auth-oauthlib = "^1.2.2"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.174.0"
gspread = "^6.2.1"
yt-dlp = "^2025.6.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
